Update sell menu
