+++ plugins/shared/src/main/kotlin/fr/pickaria/shared/PlayerExtensions.kt
@@ -1,6 +1,8 @@
 package fr.pickaria.shared
 
+import org.bukkit.Material
 import org.bukkit.entity.HumanEntity
+import org.bukkit.inventory.Inventory
 import org.bukkit.inventory.ItemStack
 
 fun HumanEntity.give(vararg items: ItemStack) {
@@ -10,3 +12,5 @@ fun HumanEntity.give(vararg items: ItemStack) {
 		item.velocity = location.direction.multiply(0.3)
 	}
 }
+
+fun Inventory.count(material: Material) = contents.filterNotNull().filter { it.type == material }.sumOf { it.amount }
